council/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
langgraph==0.2.45
langchain-core==0.3.15
httpx==0.27.2
redis==5.2.0
pyyaml==6.0.2
